+++ src/candle/config/persistence/xml/xmlpersister.h
@@ -13,7 +13,7 @@ class XmlPersister : public Persister
         void close() override;
         bool setInt(const QString group, const QString key, const int value) override;
         bool setString(const QString group, const QString key, const QString value) override;
-        bool setFloat(const QString group, const QString key, const float value) override;
+        bool setDouble(const QString group, const QString key, const double value) override;
         bool setBool(const QString group, const QString key, const bool value) override;
 
     private:
